Move executable_pex_tool for sane deps.
