+++ src/python/pants/python/setup_py_runner.py
@@ -9,7 +9,7 @@ from pex.interpreter import PythonInterpreter
 from pex.pex import PEX
 from pex.pex_info import PexInfo
 
-from pants.backend.python.subsystems.executable_pex_tool import ExecutablePexTool
+from pants.python.executable_pex_tool import ExecutablePexTool
 from pants.option.optionable import Optionable
 from pants.python.python_requirement import PythonRequirement
 from pants.util.contextutil import pushd
